+++ src/Pages/EditArticlePage/Components/index.jsx
@@ -1,0 +1,3 @@
+import EditArticlePageBody from './EditArticlePageBody';
+
+export { EditArticlePageBody };
