--- src/Pages/EditArticlePage/Conponents/index.jsx
@@ -1,4 +1,0 @@
-import EditArticlePageHeader from './EditArticlePageHeader';
-import EditArticlePageBody from './EditArticlePageBody';
-
-export { EditArticlePageHeader, EditArticlePageBody };
